STYLE: Major refactoring
